Allow styling of image and title tags
Move the `jk-amazon-image` and `jk-amazon-info-title` classes into the
`image` and `title` methods instead of using them on wrapper divs in
the `detail` method.

This allows them to be styled even when using them on their own.

Note that it might be desirable to un-nest the styles in
`jk_amazon.scss` so that the tags have their default styling even when
used on their own.  But the `padding-bottom` for the title is likely
not wanted outside the context of the `detail` tag.